refactor: Zustand 스토어 구현 변경 및 OrderStoreProvider 제거
