+++ src/app/layout.tsx
@@ -1,5 +1,4 @@
 import NavigationBar from "@/components/ui/NavigationBar";
-import { OrderStoreProvider } from "@/provider/order-store-provider";
 import { ReactQueryClientProvider } from "@/provider/ReactQueryClientProvider";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import type { Metadata } from "next";
@@ -13,7 +12,7 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
@@ -26,10 +25,8 @@ export default function RootLayout({
       >
         <ReactQueryClientProvider>
           <ReactQueryDevtools initialIsOpen={false} />
-          <OrderStoreProvider>
-            <NavigationBar />
-            {children}
-          </OrderStoreProvider>
+          <NavigationBar />
+          {children}
         </ReactQueryClientProvider>
       </body>
     </html>
